Fixing some bugs in the API data retrieval code

+++ app/src/main/java/br/com/marcottc/weatherpeek/model/CurrentWeatherData.kt
@@ -13,12 +13,12 @@ data class CurrentWeatherData(
     @SerializedName("dew_point") var dewPoint: Double,
     var clouds: Int,
     var visibility: Int,
-    @SerializedName("wind_speed") var windSpeed: Int,
+    @SerializedName("wind_speed") var windSpeed: Double,
     @SerializedName("wind_deg") var windDeg: Double,
     @SerializedName("weather") var weatherList: List<WeatherData>,
     var rain: RainData
 ) {
     constructor(dt: Long, sunrise: Long, sunset: Long, temp: Double, pressure: Int, humidity: Int, clouds: Int) :
-        this(dt, sunrise, sunset, temp, 0.0, pressure, humidity, 0.0, clouds, 0, 0, 0.0,
+        this(dt, sunrise, sunset, temp, 0.0, pressure, humidity, 0.0, clouds, 0, 0.0, 0.0,
         listOf(WeatherData(304, "Sunny", "Clear sunny", "08n")), RainData(0.0))
 }
